Move Packet from core::packet to sdk::packet
This way we don't need to split perf_libs yet.
